Finalized Sequential Progress Estimation Algorithm
It turned out on line 252, when a better best cost is found before
a node is taken from the enumeration list, that I was not previously
actually tracking that as progress made, which obviously led to several
cases of ridiculously low progress values on solved instances.

Removed (only commented out if later important) all the remainder-checking
based on remaining_share, which only ever appeared to make a difference
of approximately 10^-12 at most, a portion easily ignored. It did, however,
have a severe impact on the effect the overall progress estimator had on
runtime, appearing to increase the magnitude of this impact from less
than 2% to rarely as much as 10%. With other related optimizations,
the progress estimation has returned to being roughly insignificant in
comparison to the total runtime complexity of the algorithm.

Added an internal debugging tool for the progress estimation algorithm
itself, based on local_progress_accumulator. This may be useful, as it
was when developing this algorithm, so I haven't completely removed it,
but left it commented out for now since it is not important work and
can represent a small impact on runtime.

Added a compiler-time constant ENABLE_PROGRESS_ESTIMATION and gated all
progress estimation work behind that constant resolving TRUE.

Cleaned up vestigial code sections from this part of the project, such
as the get_estimated_progress_percent function which is no longer
necessary at all.