Converted progress estimation to work with the parallel solver
Progress is tracked just like in sequential, by an integer out of
ULLONG_MAX, seperately in each thread tracked by the estimated_trimmed_percent
array, indexed by thread_id.

Because of imprecision, the total progress on a completed instance
is not expected to reach exactly ULLONG_MAX, but the difference is
expected to be much less than 10^-5.

There is still some problem, either in some part of work stealing or
thread restart, that leads to instances sometimes reporting only
99% progress even when finished, but this appears to be rare and
probably not very important for our intended purpose.